// WADNR.Web/src/app/shared/generated/model/invoice-upsert-request.ts
export class InvoiceUpsertRequest { 
    InvoicePaymentRequestID: number;
    InvoiceNumber: string;
    InvoiceIdentifyingName?: string | null;
    InvoiceDate: string;
    PaymentAmount?: number | null;
    MatchAmount?: number | null;
    InvoiceMatchAmountTypeID: number;
    FundSourceID?: number | null;
    Fund?: string | null;
    Appn?: string | null;
    SubObject?: string | null;
    ProgramIndexID?: number | null;
    ProjectCodeID?: number | null;
    OrganizationCodeID?: number | null;
    InvoiceStatusID: number;
    InvoiceApprovalStatusID: number;
    InvoiceApprovalStatusComment?: string | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}