// WADNR.Web/src/app/shared/generated/model/invoice-payment-request-upsert-request.ts
/**
 * WADNR.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class InvoicePaymentRequestUpsertRequest { 
    ProjectID: number;
    VendorID?: number | null;
    PreparedByPersonID: number;
    InvoicePaymentRequestDate: string;
    PurchaseAuthority?: string | null;
    PurchaseAuthorityIsLandownerCostShareAgreement?: boolean;
    Duns?: string | null;
    Notes?: string | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface InvoicePaymentRequestUpsertRequestForm { 
    ProjectID: FormControl<number>;
    VendorID?: FormControl<number>;
    PreparedByPersonID: FormControl<number>;
    InvoicePaymentRequestDate: FormControl<string>;
    PurchaseAuthority?: FormControl<string>;
    PurchaseAuthorityIsLandownerCostShareAgreement?: FormControl<boolean>;
    Duns?: FormControl<string>;
    Notes?: FormControl<string>;
}

export class InvoicePaymentRequestUpsertRequestFormControls { 
    public static ProjectID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: true,
            validators: 
            [
                Validators.required,
            ],
        }
    );
    public static VendorID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static PreparedByPersonID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: true,
            validators: 
            [
                Validators.required,
            ],
        }
    );
    public static InvoicePaymentRequestDate = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: true,
            validators: 
            [
                Validators.required,
            ],
        }
    );
    public static PurchaseAuthority = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
                Validators.maxLength(255),
            ],
        }
    );
    public static PurchaseAuthorityIsLandownerCostShareAgreement = (value: FormControlState<boolean> | boolean = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<boolean>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static Duns = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
                Validators.maxLength(20),
            ],
        }
    );
    public static Notes = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}